[4기 - 황창현] Week2-4 JdbcTemplate을 활용한 바우처 지갑 CREATE, READ, DELETE 기능 구현

+++ src/main/java/com/programmers/springweekly/service/CustomerService.java
@@ -7,13 +7,13 @@ import com.programmers.springweekly.dto.customer.response.CustomerListResponse;
 import com.programmers.springweekly.dto.customer.response.CustomerResponse;
 import com.programmers.springweekly.repository.customer.CustomerRepository;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.UUID;
-import java.util.stream.Collectors;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class CustomerService {
 
     private final CustomerRepository customerRepository;
@@ -41,25 +41,36 @@ public class CustomerService {
     }
 
     public CustomerResponse findById(UUID customerId) {
-        Customer customer = customerRepository.findById(customerId);
+        Customer customer = customerRepository.findById(customerId)
+                .orElseThrow(() -> new NoSuchElementException("찾는 고객이 없습니다, 고객이 저장되어있는지 확인해보세요."));
+
         return new CustomerResponse(customer);
     }
 
     public CustomerListResponse findAll() {
         List<Customer> customerList = customerRepository.findAll();
-        return new CustomerListResponse(customerList.stream().map(CustomerResponse::new).collect(Collectors.toList()));
+        return new CustomerListResponse(customerList.stream().map(CustomerResponse::new).toList());
     }
 
     public CustomerListResponse getBlackList() {
         List<Customer> customerList = customerRepository.getBlackList();
-        return new CustomerListResponse(customerList.stream().map(CustomerResponse::new).collect(Collectors.toList()));
+        return new CustomerListResponse(customerList.stream().map(CustomerResponse::new).toList());
     }
 
-    public void deleteById(UUID customerId) {
-        customerRepository.deleteById(customerId);
+    public int deleteById(UUID customerId) {
+        if (!customerRepository.existById(customerId)) {
+            throw new NoSuchElementException("삭제하려고 입력한 고객의 ID는 존재하지 않는 ID입니다. 다시 입력해주세요");
+        }
+
+        return customerRepository.deleteById(customerId);
     }
 
     public void deleteAll() {
         customerRepository.deleteAll();
     }
+
+    public boolean existById(UUID customerId) {
+        return customerRepository.existById(customerId);
+    }
+
 }
